refactor - break-damage-form: Split component
* Create break-damage-form.component
* Move implementation of break damage to break-damage-form
* Fix up styling
