+++ src/app/app.module.ts
@@ -4,12 +4,14 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 
 import { AppComponent } from '@components/app/app.component';
+import { BreakDamageFormComponent } from '@components/break-damage-form/break-damage-form.component';
 
 import { AppService } from '@services/app.service';
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    BreakDamageFormComponent
   ],
   imports: [
     BrowserModule,
