+++ src/app/components/app/app.component.html
@@ -1,77 +1,5 @@
 
 <div class="container">
     <h1>Mobius Final Fantasy Toolkit v{{version}}</h1>
-    <form [formGroup]="breakDamageForm" novalidate>
-        <legend>Break Damage Calculation</legend>
-        <fieldset class="form-group">
-            <h5>Modifiers</h5>
-            <div class="form-check form-check-inline">
-                <label class="form-check-label">
-                    <input class="form-check-input" type="checkbox" formControlName="hasBoost" value="" />
-                    Boost
-                </label>
-            </div>
-            <div class="form-check form-check-inline">
-                <label class="form-check-label">
-                    <input class="form-check-input" type="checkbox" formControlName="hasTrance" value="" />
-                    Trance
-                </label>
-            </div>
-            <div class="form-check form-check-inline">
-                <label class="form-check-label">
-                    <input class="form-check-input" type="checkbox" formControlName="hasBDD" value="" />
-                    Break Defense Down
-                </label>
-            </div>
-            <div class="form-check form-check-inline">
-                <label class="form-check-label">
-                    <input class="form-check-input" type="checkbox" formControlName="hasEnElement" value="" />
-                    EnElement
-                </label>
-            </div>
-            <div class="form-check form-check-inline">
-                <label class="form-check-label">
-                    <input class="form-check-input" type="checkbox" formControlName="hasWeaken" value="" />
-                    Weaken
-                </label>
-            </div>
-        </fieldset>
-
-        <fieldset class="form-group">
-            <div class="form-row">
-                <div class="col-md">
-                    <label for="inputBaseBreak">Base Break Power</label>
-                    <input type="number" class="form-control" id="inputBaseBreak" formControlName="baseBreak" />
-                </div>
-                <div class="col-md">
-                    <label for="inputExploitWeakness">Exploit Weakness</label>
-                    <input type="number" class="form-control" id="inputExploitWeakness" formControlName="exploitWeakness" />
-                </div>
-                <div class="col-md">
-                    <label for="inputPiercingBreak">Piercing Break</label>
-                    <input type="number" class="form-control" id="inputPiercingBreak" formControlName="piercingBreak" />
-                </div>
-            </div>
-        </fieldset>
-
-        <fieldset class="form-group">
-            <h5>Total Break Power: {{ totalBreakPower }}</h5>
-        </fieldset>
-
-        <fieldset class="form-group">
-            <label for="inputTargetBreakGauge">Target's Break Gauge</label>
-            <input type="number" class="form-control" id="inputTargetBreakGauge" formControlName="targetBreakGauge" />
-        </fieldset>
-
-        <legend>Break Reduction Calculation</legend>
-
-        <fieldset class="form-group">
-            <legend>Current Form Value:</legend>
-            <pre>
-                <code>{{ testForm.value | json }}</code>
-            </pre>
-            <h5>Current Form Status: {{ testForm.status | json }}</h5>
-
-        </fieldset>
-    </form>
+    <app-break-damage-form></app-break-damage-form>
 </div>
