+++ src/app/components/break-damage-form/break-damage-form.component.html
@@ -1,0 +1,77 @@
+<form [formGroup]="breakDamageForm" class="card" novalidate>
+    <legend class="card-header">Break Damage Calculation</legend>
+    <div class="list-group list-group-flush">
+        <div class="list-group-item">
+            <fieldset class="form-group">
+                <h5>Modifiers</h5>
+                <div class="form-check form-check-inline">
+                    <label class="form-check-label">
+                        <input class="form-check-input" type="checkbox" formControlName="hasBoost" value="" />
+                        Boost
+                    </label>
+                </div>
+                <div class="form-check form-check-inline">
+                    <label class="form-check-label">
+                        <input class="form-check-input" type="checkbox" formControlName="hasTrance" value="" />
+                        Trance
+                    </label>
+                </div>
+                <div class="form-check form-check-inline">
+                    <label class="form-check-label">
+                        <input class="form-check-input" type="checkbox" formControlName="hasBDD" value="" />
+                        Break Defense Down
+                    </label>
+                </div>
+                <div class="form-check form-check-inline">
+                    <label class="form-check-label">
+                        <input class="form-check-input" type="checkbox" formControlName="hasEnElement" value="" />
+                        EnElement
+                    </label>
+                </div>
+                <div class="form-check form-check-inline">
+                    <label class="form-check-label">
+                        <input class="form-check-input" type="checkbox" formControlName="hasWeaken" value="" />
+                        Weaken
+                    </label>
+                </div>
+            </fieldset>
+
+            <fieldset class="form-group">
+                <h5>Auto-Abilities</h5>
+                <div class="form-row">
+                    <div class="col-md">
+                        <label for="inputBaseBreak">Base Break Power</label>
+                        <input type="number" class="form-control" id="inputBaseBreak" formControlName="baseBreak" />
+                    </div>
+                    <div class="col-md">
+                        <label for="inputExploitWeakness">Exploit Weakness</label>
+                        <input type="number" class="form-control" id="inputExploitWeakness" formControlName="exploitWeakness" />
+                    </div>
+                    <div class="col-md">
+                        <label for="inputPiercingBreak">Piercing Break</label>
+                        <input type="number" class="form-control" id="inputPiercingBreak" formControlName="piercingBreak" />
+                    </div>
+                </div>
+            </fieldset>
+        </div>
+        <div class="list-group-item">
+            <h5 class="mb-0">Total Break Power: {{ totalBreakPower }}</h5>
+        </div>
+        <div class="list-group-item">
+            <fieldset class="form-group">
+                <label for="inputTargetBreakGauge">Target's Break Gauge</label>
+                <input type="number" class="form-control" id="inputTargetBreakGauge" formControlName="targetBreakGauge" />
+            </fieldset>
+        </div>
+    </div>
+    <div class="card-footer">
+        <button class="btn btn-light" data-toggle="collapse" data-target=".debug-info">Show Debug Info</button>
+        <fieldset class="form-group debug-info mt-3 collapse">
+            <h5>Current Form Value:</h5>
+            <pre>
+                <code>{{ breakDamageForm.value | json }}</code>
+            </pre>
+            <h5>Current Form Status: {{ breakDamageForm.status | json }}</h5>
+        </fieldset>
+    </div>
+</form>
